// src/app/types/buttons.ts
export type LinkButtonSettings = {
    type: 'link';
    /** Label displayed inside the button */
    text: string;
    /** Button decoration (icon, arrow) or left blank for none */
    decoration?: BtnDecoration;
    /** React Router path or absolute URL the button links to */
    link: string;
    /** Whether the link opens in a new tab @default '_self' */
    target?: "_blank" | "_self";
}
// ...
export type BtnArrowSettings = {
    type: 'arrow';
    arrowSide?: "left" | "right";
    arrowDirection?: ArrowDirectionProps;
};

type BtnIconSettings = {
    type: "icon";
    icon: Icon;
}

export type BtnDecoration = BtnArrowSettings | BtnIconSettings;